# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api-rbac_internal_model.AssignRolesRequest:
    properties:
      role_ids:
        description: '例: [1, 2]'
        items:
          type: integer
        type: array
    required:
    - role_ids
    type: object
  api-rbac_internal_model.ChangePasswordRequest:
    properties:
      new_password:
        example: new123456
        maxLength: 128
        minLength: 6
        type: string
      old_password:
        example: old123456
        maxLength: 128
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  api-rbac_internal_model.CreateUserRequest:
    properties:
      email:
        example: zhangsan@example.com
        maxLength: 128
        type: string
      password:
        example: "123456"
        maxLength: 128
        minLength: 6
        type: string
      username:
        example: zhangsan
        maxLength: 64
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
...
  api-rbac_internal_model.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
      updated_at:
        type: string
    type: object
  api-rbac_internal_model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/api-rbac_internal_model.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  api-rbac_internal_model.UpdateUserRequest:
    properties:
      email:
        example: newemail@example.com
        maxLength: 128
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
    type: object
  api-rbac_internal_model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          $ref: '#/definitions/api-rbac_internal_model.Role'
        type: array
      status:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  api-rbac_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  api-rbac_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: 通用的 RBAC 权限管理微服务，支持用户/角色/权限的 CRUD 及绑定关系管理。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RBAC 权限管理系统 API
  version: "1.0"
paths:
...
  /users:
    get:
      description: 分页查询用户列表，支持关键词搜索
      parameters:
      - default: 1
        description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        example: 10
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词(用户名/邮箱)
        example: admin
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_pkg_response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 新增一个用户
      parameters:
      - description: 创建用户参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_internal_model.User'
              type: object
        "400":
          description: 参数错误或用户名已存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      description: 根据 ID 删除一个用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      description: 根据 ID 获取用户信息，包含关联的角色和权限
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_internal_model.User'
              type: object
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新用户的邮箱和状态
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - 用户管理
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: 修改指定用户的密码，需提供旧密码验证
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 修改密码参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "400":
          description: 旧密码错误
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - 用户管理
  /users/{id}/roles:
    post:
      consumes:
      - application/json
      description: 覆盖式更新用户的角色列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.AssignRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 为用户分配角色
      tags:
      - 用户管理
  /users/{id}/roles/{roleId}:
    delete:
      description: 移除用户的某个角色
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 移除用户角色
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token} 格式的 JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"